+++ tests/integration/Assets/Packages/Package/SubPackages/SubPackage/Controllers/ExampleController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Integration\Assets\Packages\Package\SubPackages\SubPackage\Controllers;
+
+/**
+ * @author Alex Ash <streamprop@gmail.com>
+ */
+class ExampleController
+{
+    public function getAction($alpha, $betta)
+    {
+        return "Integration\\Assets\\Packages\\Package\\SubPackages\\SubPackage\\Controllers\\ExampleController::getAction($alpha, $betta)";
+    }
+}