+++ CHANGELOG.md
@@ -11,3 +11,4 @@ All notable changes to this project will be documented in this file.
 - Getting router via DI container
 - Guessing controller namespace by URI path
 - Guessing controller namespace with multi-word uri path segments
+- Guessing controller namespace through hierarchy of packages